docker_kryptos/kryptos/requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2
django-cors-headers==4.6.0
django-rest-framework-mongoengine==3.4.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
mongoengine==0.29.1
multidict==6.1.0
propcache==0.2.1
PyJWT==2.10.1
pymongo==4.10.1
sqlparse==0.5.3
urllib3==2.3.0
yarl==1.18.3
